Port anitya to use urllib2 instead of python-requests
The problem of python-requests is that it does not support calls to ftp
servers. We have a number of projects using FTP and providing FTP urls.

The problem with urllib and urllib2 is that neither actually check if
the SSL certificate is valid or matching the hostname.
Unfortunately, we'll have to do without for two reasons: a) too many
projects are FTP based and b) there is no option in the UI (for the
moment) to disable that SSL check.