Add a way to find the projects that simply never updated
Sometimes projects fail to update, but they did update before, it's just
that at the last cron run the server was down or something like this.
But we also have projects that simply never have been able to update, not
even once. Maybe the info we have about these projects are simply wrong,
maybe the project doesn't exists anymore.

This commit allows to find this type of projects