Allow an initial mapping when creating a project.
When I'm adding an upstream project for the very first time, I want this quite
often -- I want to map packages in Fedora.  When I have 2 trillion packages to
map, it can be quite annoying without this: I have to open the form the create
the upstream project, submit it, click 'add a mapping', then fill out that
form, then submit it.. over and over for every project/package.

This commit alters that first 'new upstream project' form and gives it a
section to add an initial distro mapping in one go.

Fixes #52.