Add api to query for a project by homepage.
I'm thinking that we can use this when a new package gets added to Fedora.  We
can look up its upstream_url in pkgdb, and then query anitya to see if any
projects are registered with that homepage.  If there are not, then add one
with this new package if its backend is an easy one (pypi, rubygems, etc).