Reset the justedit session variable as soon as it is set
But we still forward it to the template so that the template can make use
of it.
In this case, since the edit function in the UI sends a redirect at the
end, the session variable will only be set during the redirection from
edit project to view project (where it is made None again).
Since should ensure the justedit session variable cannot be used accross
requests