Invert the name and SF name in the SourceForge backend.
Until now we were using the name field as the name of the project on SF,
so sf.net/<project>. But this causes problems when the project releases
a tarball with a different name than their project name, so we introduced
the SF name, aimed to be the name of the tarball on SF.
However, this does not cover the projects releasing multiple tarball.

In anitya project and homepage are required to be unique, so a project
releasing multiple tarballs under the same project/homepage was a
problem.

By inverting the project name and the SF name we can solve this.

From now on: the project name is the name of the tarball to search in the
RSS page and the SF name is the name of the project on SourceForge.

Fixes https://github.com/fedora-infra/anitya/issues/80
Fixes https://github.com/fedora-infra/anitya/issues/146