Adjust the rubygems backend to raise an exception if the version retrieved is unknown
In commit a099bb7a0dca1e01c68feaa3f94c1c78a87ef495 we changed the endpoint
used to retrieve rubygems versions.
This new endpoint returns a JSON blob even if the package does not exists,
unlike the previous endpoint.

So we should check the version retrieved and raise an exception if this version
is unknown.