Handle projects hosted on github without a version_url
Lots of projects hosted on github are setting a homepage but no
version_url, this makes the cron fail to retrieve their information.
With this change, if there is not version_url but the homepage is on
github, we'll use the homepage to generate the url used to retrieved
the tags/versions.

Adjust the unit-tests to cover this case