Report the reason fetching a url failed in the folder backend
It was noted in https://github.com/fedora-infra/anitya/issues/338 that
FTP urls fail. In my tests, FTP connections to the host mentioned in the
issue fail occasionally when the connection is reset by the host. This
does not make the FTP call more robust, but it does report the actual
error that occurred.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>