Projects with the folders backend can use insecure HTTPS
Projects have a flag to allow for insecure HTTPS connections where the
signature on the certificate presented by the server is not verified.
This enables users to set that flag when creating or editing projects
using the folder backend. The reason for this is that the backend URL
can be any page on the Internet and there's no guarantee that user is
serving a certificate signed by a trusted CA (or they might have messed
up their configuration).

fixes #386

Signed-off-by: Jeremy Cline <jeremy@jcline.org>