Do not install system versions of Python dependencies
Currently, we have to set up the virtualenv with the system site
packages since we have a hard dependency on the rpm module which isn't
on PyPI (as far as I can tell).

Although it is nice to have the system version of packages available
outside the virtual environment for testing, they cause problems when
added to the virtual environment. Specifically, it caused the flask
binary to not be inside the virtualenv and that is used to run the web
application. This commit removes all the non-essential system packages
to ensure the virtualenv is clean.

fixes #415

Signed-off-by: Jeremy Cline <jeremy@jcline.org>