Decouple the Anitya config from Flask
This commit changes how the configuration for Anitya is loaded. The
motivation for this is to get rid of the troubling circular imports
caused by the backend using a scoped session that's created by the
front-end. This is done because the configuration is loaded by Flask.

Now, the configuration is loaded in the ``anitya.config`` module. This
module has no dependencies on other pieces of Anitya. This will let the
backend set up the scoped session. The Flask frontend, which (obviously)
already depends on the backend can then make use of that scoped session.

Note: I changed the configuration format to YAML (because I like YAML),
but if people are opposed to this it can be easily switched back to a
Flask-compatible Python file. I _also_ removed the ANITYA_LOG_LEVEL
setting in favor of using a dict config.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>