Remove the old docs directory.
This removes all the machinery for building the RST into HTML, updates
the templates to point to the static docs built from the Sphinx project,
and adds redirects for the old routes.

This configures Flask to serve the static documentation, but when
deployed behind a web server like Apache httpd or nginx the user can
easily rewrite URLs and serve the static files directly.

fixes #454

Signed-off-by: Jeremy Cline <jeremy@jcline.org>