Stop using pkg_resources
This drops all the various pkg_resources statements. Firstly, we no
longer support Python 2.6 and the statements were to support installing
Anitya on EL6 which no longer works since we don't support 2.6.

Secondly, this is a downstream packaging concern that I would prefer to
carry as a patch. It causes all sorts of problems. For example, having
it in the setup.py means you must have SQLAlchemy installed before you
can install Anitya which is annoying.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>