Remove the explicit dependency on Flask to avoid setuptools bug
At the moment you cannot ``python setup.py install`` Anitya (and this is
causing our documentation to not build) because we're hitting a
setuptools bug[0] which is causing "Flask-WTF" to be matched for the
"Flask" requirement. Without the explicit requirement, the Flask plugins
we use (including Flask-WTF) will pull Flask in.

[0] https://github.com/pypa/setuptools/issues/196

Signed-off-by: Jeremy Cline <jeremy@jcline.org>