Issue #533: sort all projects by (name, ecosystem)
The list of all projects is already sorted by the project name. But it
results in inconsistent order over long periods of time (~8 hours) as
there is no explicit rule to sort two projects with the same name.
Two examples of multiple projects with the same name are readline and
zlib.

For api v2, add a second sort rule, based on the ecosystem, to make the
results predictable.

Signed-off-by: Ricardo Martincoski <ricardo.martincoski@gmail.com>