Provide a more helpful error to users of multiple providers
If a user signs in with identity provider A and then signs in with
identity provider B and both providers have the same email for that
user, an IntegrityError happens when social_auth tries to make a new
user. Handle this and inform the user what provider they should use to
authenticate.

Ideally we could merge the accounts and let users auth with either one,
but this is more involved because not all providers confirm the email of
the user, so if a malicious user signs up with another user's email and
then it got merged, they could perform actions as the user.

Fixes #559

Signed-off-by: Jeremy Cline <jcline@redhat.com>