Simplify logs
This commit is introduces simple status for every project instead of
current robust log table.
It drops the current log table and adds check_successfull flag to
project.
The frontend page for logs is refactored to work with this change.

This change should make working with logs simpler and remove unwanted
table from database.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>