Fix first login issue
This is fixing the issue that was accidentally added by
social_auth check in events.
It turned out that the user table entry and social_auth
entry are created in separate transactions.
This caused issue when you tried to login when user didn't
exists.
This fix is checking the social_auth entry when user logs
in and deletes the user when no entry is actually created.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>