Versions are now shown correctly on project page
With introduction of new version prefix handling a new issue was
created. With the new prefix handling we actually don't remove the
prefix, but instead we let this on Version class.
Previously the `project.html` template used version.version to show
version on the project page, but now it is changed to `version.parse()`
which returns the correct string stripped by any prefix.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>