Fix: Comparing by dates created version duplicates
Date version scheme compared for equality only by dates, this caused
creation of duplicities, which led to crash when commiting changes.
This commit changes the way how the date version scheme is comparing for
equality. Now it is calling __eq__ method in Version class. Versions are
now considered equal if they are same after prefix is stripped.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>