Use SSE consumer instead of fedmsg consumer for Libraries.io
Because the libraries.io SSE stream is used only by Anitya. It is a
logical move to make it part of Anitya itself.

This PR is changing the current libraries.io consumer, which
listened to fedmsg messages emitted by sse2fedmsg, to SSE client
which is listening directly to libraries.io SSE stream.

This PR also removes last fedmsg related code. So we can consider Anitya
to be Fedora Messaging ready. \o/

Signed-off-by: Michal Konečný <mkonecny@redhat.com>