Make prepare_query() easier to understand
I had trouble understanding how prepare_query() builds GraphQL queries,
in part because some braces weren't opened and closed on the same level.
The resulting queries change a little in whitespace (some elements are
on the same line now that were on separate ones previously), but this
doesn't change the semantics.

Signed-off-by: Nils Philippsen <nils@redhat.com>