Add error counter for project checking
This adds error counter to project, which is incremented each time a
error happens when checking for new version. Error counter is reset to
0, when succesful check is done.

This also introduces GUI changes that allows users to see the error
counter on each project and option to find the projects, which are
failing most.

This will help identify problematic projects.

Implements part of #829

Signed-off-by: Michal Konečný <mkonecny@redhat.com>