Use mapping as another condition for deletion
Recently there was a flood of junk projects from PyPi received through
libraries.io consumers. These projects doesn't exist anymore on PyPi,
but at least one version was retrieved, so current solution of delete
candidate recognition doesn't recognize this project as candidate
for deletion, but if the project doesn't have mapping it could be
in most cases safely deleted.

Still if there is mapping and no versionis retrieved it will be deleted.
The prime condition is still the error threshold.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>