Fix internal server error on some projects with RPM version scheme
This error was caused by nonsense version which unfortunately matched the rpm
regex. This wasn't issue before, because we used our own implementation of
compare method, but now we can use the one from RPM library.

The solution to this was to make sure that the version we are trying to compare
is not an empty string.

Second thing that was part of this internal server error issue and caused a lot
of mess in whole infra are the hanging connection to database when the above
error occurs. So this commit makes sure that the session is always during
closed check_service execution.

Fixes #1390.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>