Fix handling of JSON requests in APIv2
When specifying the header as `application/json` in POST request the flask
parser wasn't able to parse the request correctly and instead saw empty list and
returned 400.

This fix will check if the request is JSON request and if this is true it will
look for the data in request.json instead of request.form.

Fixes #1464

Signed-off-by: Michal Konečný <mkonecny@redhat.com>