Improve the dependency management for javascript dependencies
This commit will add npm for dependency management of javascript dependencies.
Most of the static folder was restructured.

Unfortunately this updates some of the packages to newer versions, which needs
templates adjusted. Because of this Bootstrap 3 -> Bootstrap 5 migration is part
of this commit.

As part of the Bootstrap migration there is a cleanup of unused templates.

It also updates documentation to reflect this new change and updates both dev
environments to use the npm installed packages.

Closes #1504

Signed-off-by: Michal Konečný <mkonecny@redhat.com>