Support odd-numbered version filtering
This commit introduces a filter pattern `!odds[:n]`, which can be used
to help identify pre-release versions for packages using odd-numbered
versions for development releases. This convenience filter can be used
over having to add explicit exceptions for each odd version entry for
projects using odd-numbered development releases.

For example, if a project configures a pre-release filter using the
value `!odds`, any odd-numbered minor version for a project will be
flagged as a pre-release. For example, versions `1.0` and `1.2` will be
considered stable versions, as where `1.1` is considered a pre-release
version.

Resolves: #1088
Signed-off-by: James Knight <james.d.knight@live.com>