Remove straight.plugin from anitya dependencies
This library caused issue with python 3.12 and as there is no update to this
library for 3 years it is best to just remove it.

The new solution is using standard libraries instead and doesn't introduce new
dependencies. Also no test was touched, so the logic is still the same.

Fixes #1769

Signed-off-by: Michal Konecny <mkonecny@redhat.com>