Fix: Internal server error when adding packages
When adding packages sometimes the query finds more than one entry.
Although there are no duplicates in the database it happens sometimes.
To prevent this let's pick the first entry from the returned results.

This is not the final fix, but it should be enough to resolve the issue
with 500 for now.

Fixes #1911