Add check for version_url to API
The API doesn't check if the version_url is filled when the backend
requires it. The project created this way will be removed after 1000
failed checks, but we still should prevent this by checking if the
version_url is filled.

So this PR is adding new attribute to backends, which is checked when
creating new project using API.

Fixes #1921