Check GitHub ratelimit correctly
Just removing the ratelimit threshold will cause Anitya to fail on
KeyError as instead of 403 when the ratelimit is reached GitHub returns
200 and json with error structure, which is completely different then
standard response. To fix this we need to check the response header for
specific parameters regarding rate limit.

This change is checking X-RateLimit-Remaining and X-RateLimit-Reset
parameters in response header and raising RateLimitException when the
limit is reached.

Signed-off-by: Michal Konecny <mkonecny@redhat.com>